[*] Разбила функцию Screen на отдельные компонентные функции Header Logo и ListHero
